refactor(cli): extract EngineRecipe boot module from main
Move shared engine construction (punct/ITN/VAD/hotwords/pool/threads)
into crates/gigastt/src/boot.rs so serve (incl. admin reload) and
offline transcribe/batch/watch share one recipe. CLI flags unchanged.
