+++ crates/gigastt/src/lib.rs
@@ -1,3 +1,4 @@
 pub use gigastt_core::*;
 pub mod batch;
+pub mod boot;
 pub mod server;
